Introduce ldap_ping_test option
ldap_ping_test allows to turn :on or :off the LDAP Server availability
check, which performs a simple ping (ICMP) test.

This fails in networks for LDAP servers which filter or drop ICMP
packages. (This is pretty rare but we hit this in at least one
particular customer environment).

Also enhanced the description of the max_ldap_attempts option, to point
out that those connect attempts are just plain ping attempts.

Conflicts:

	src/api/app/models/user.rb
	src/api/config/environments/production.rb